TestNG_HM в процессе

+++ src/main/java/Calculator.java
@@ -8,7 +8,7 @@ public class Calculator {
     public int sum(int a, int b){ return a + b;}
     public int sub(int a, int b){ return a - b;}
     public int multi(int a, int b){ return a * b;}
-    public double div(int a, int b){ return a / b;}
+    public double div(double a, double b){ return a / b;}
 
     public String getTitle() {return title;}
 }
